requirements.txt
asgiref==3.8.1
colorama==0.4.6
coverage==7.9.2
dj-database-url==3.0.1
Django==5.2.1
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-nested-routers==0.94.2
environ==1.0
Faker==37.3.0
flake8==7.3.0
greenlet==3.2.3
gunicorn==23.0.0
iniconfig==2.1.0
mccabe==0.7.0
mysql==0.0.3
mysql-connector-python==9.3.0
mysqlclient==2.2.7
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
psycopg2-binary==2.9.10
pycodestyle==2.14.0
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.9.0
PyMySQL==1.1.1
pytest==8.4.1
pytest-cov==6.2.1
pytest-django==4.11.1
python-decouple==3.8
SQLAlchemy==2.0.41
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
whitenoise==6.9.0
